Update README.md
In this study, we plan to use M-
Qubed, an algorithm that can make profitable compromises, in an economic model.
Our economic model is based on the Price-Quality model. We want to test whether an
M-Qubed agent can maximize profit while avoiding myopic pricing strategies. Then,
in our model, a Genetic Algorithm is used to determine the optimal parameters for an
M-Qubed agent.